% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_weight_matrix2.R
\name{cells_weight_matrix2}
\alias{cells_weight_matrix2}
\title{Day hello}
\usage{
cells_weight_matrix2(coords, groups, labels, alpha = 1)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
