% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relevant_genes.R
\name{get_relevant_genes}
\alias{get_relevant_genes}
\title{Identify the relevant genes from wSIR}
\usage{
get_relevant_genes(out, p.threshold = 0.01)
}
\arguments{
\item{out}{output from running wSIR function}

\item{p.threshold}{threshold value for unadjusted p-value}
}
\value{
character vector of relevant features
}
\description{
This function calculates standardized z-scores for gene expression values, converts them to corresponding p-values, identifies genes with p-values below a specified threshold, and returns their names as relevant genes.
}
\examples{
#hello()

}
