% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplier.R
\name{multiplier}
\alias{multiplier}
\title{Day hello}
\usage{
multiplier(data, pc_dirs)
}
\arguments{
\item{data}{no arguments}

\item{pc_dirs}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
#hello()

}
