% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{plot_umap}
\usage{
plot_umap(exprs, umap_coords, highest_genes = NULL, genes = NULL, n_genes, ...)
}
\arguments{
\item{exprs}{matrix containing normalised gene expression data including n cells and p genes, dimension n * p.}

\item{umap_coords}{UMAP coordinates for each cell that is output of generate_umap function. The UMAP coordinates can be
based on any dimension reduction method, e.g they could be the UMAP coordinates computed on the WSIR dimension reduction
of the gene expression data, or on the PCs (principal components), or on any other low-dimensional matrix. Must be
a matrix of dimension nrow(exprs) * 2.}

\item{highest_genes}{output from top_genes function. Ensure that you only used a single WSIR direction in the top_genes
function, otherwise some genes are likely to be displayed multiple times in the resulting plots from this function.
Default is NULL so an error message can easily be thrown if genes and highest_genes are both not provided.}

\item{genes}{vector with gene names (must all be in colnames(exprs)) you wish to show in the UMAP plot. The cells
in those plots will be coloured by their expression values for the genes you provide here. Must provide either genes
or highest_genes parameter (not both): provide genes if you want to visualise a few specific genes, provide highest_genes
if you want to visualise the genes that are found to be the most important to the WSIR directions. Default is NULL
so an error message can easily be thrown if genes and highest_genes are both not provided.}

\item{n_genes}{integer for the number of genes you would like to show. Default is the number of unique genes in the
highest_genes parameter or the number of genes in the (vector) parameter genes. Use this parameter if you want to
show only a few of the most important genes (e.g select the top 4 with n_genes = 4).}

\item{...}{additional parameters for ggplot functions, e.g size (for size of points).}
}
\value{
Grid of umap plots with n_genes number of plots. Each shows the cells in a UMAP generated on the low-dimensional gene
expression data, coloured by their value for each of the genes found by top_genes.
}
\description{
A function to plot a UMAP generated on the low-dimensional embedding of the gene expression data. The points are coloured by
their value for the genes with highest (in absolute value) loading in a selected WSIR direction, by default WSIR1.
}
\examples{
data(MouseData)
wsir_obj = wSIR(X = sample1_exprs,
  coords = sample1_coords,
  optim_params = FALSE,
  alpha = 4,
  slices = 6) # create wsir object
umap_coords = generate_umap(WSIR = wsir_obj)
top_genes_obj = top_genes(WSIR = wsir_obj, highest = 4) # create top genes object
umap_plot = plot_umap(umap_coords = umap_coords,
  exprs = sample1_exprs,
  highest_genes = top_genes_obj,
  n_genes = 4)
umap_plot

}
