% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_SIR.R
\name{project_SIR}
\alias{project_SIR}
\title{Day hello}
\usage{
project_SIR(sir, newdata)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
