% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_wSIR.R
\name{project_wSIR}
\alias{project_wSIR}
\title{project_wSIR}
\usage{
project_wSIR(wsir, newdata)
}
\arguments{
\item{wsir}{wsir object that is usually the output of wSIR function. If you want to project new data into low-dim space following a
different DR method, at param wsir use a list with matrix of loadings in slot 2 (e.g PCA loadings) of dimension p * d}

\item{newdata}{matrix of new gene expression data to project into low-dimensional space. Must have the same p columns
as the columns in X argument used to generate wsir.}
}
\value{
matrix of low-dimensional representation of newdata gene expression data
}
\description{
function to project new gene expression data into low-dimensional space
}
\examples{
data(MouseData)
wsir_obj = wSIR(X = sample1_exprs, coords = sample1_coords)
sample2_low_dim_exprs = project_wSIR(wsir = wsir_obj, newdata = sample2_exprs)

}
