% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_univariate.R
\name{sir_univariate}
\alias{sir_univariate}
\title{Day hello}
\usage{
sir_univariate(
  X,
  Y,
  directions = 10,
  categorical = FALSE,
  slices = 10,
  alpha = 1,
  W
)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
