% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirPCA.R
\name{sirPCA}
\alias{sirPCA}
\title{sirPCA}
\usage{
sirPCA(
  sliced_data,
  directions,
  W = diag(nrow(sliced_data)),
  varThreshold = 0.95
)
}
\arguments{
\item{sliced_data}{matrix of scaled slice means}

\item{directions}{integer, number of directions we want in our final low-dimensional Z.}

\item{W}{matrix of slice weights. Output of createWeightMatrix function.}

\item{varThreshold}{numeric proportion of eigenvalues of variance in t(X_H) \%\emph{\% W \%}\% X_H to retain.
Default is 0.95. Select higher threshold to include more dimensions, lower threshold to include less dimensions.}
}
\value{
list containing: 1) "eigenvectors" matrix of eigenvectors of (X^H)^t \%\emph{\% W \%}\% (X^H)
2) "d" integer, selected number of dimensions
3) "evalues" vector of eigenvalues of (X^H)^t \%\emph{\% W \%}\% (X^H)
}
\description{
This function performs eigendecomposition on (X^H)^t \%\emph{\% W \%}\% (X^H), where X^H is matrix of scaled slice means.
}
\keyword{internal}
