% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicer.R
\name{slicer}
\alias{slicer}
\title{Day hello}
\usage{
slicer(X, Y, slices = 8, categorical = FALSE)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
