% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicer_categorical.R
\name{slicer_categorical}
\alias{slicer_categorical}
\title{Day hello}
\usage{
slicer_categorical(X, Y)
}
\arguments{
\item{X}{no arguments}

\item{Y}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
#hello()

}
