% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorter.R
\name{sorter}
\alias{sorter}
\title{Day hello}
\usage{
sorter(X, Y)
}
\arguments{
\item{X}{no arguments}

\item{Y}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
#hello()

}
