% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.cov.R
\name{sparse.cov}
\alias{sparse.cov}
\title{Day hello}
\usage{
sparse.cov(x)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
