% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_allocator.R
\name{spatial_allocator}
\alias{spatial_allocator}
\title{Day hello}
\usage{
spatial_allocator(coords, slices = 3)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
