% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAllocator.R
\name{spatialAllocator}
\alias{spatialAllocator}
\title{spatialAllocator}
\usage{
spatialAllocator(coords, slices = 3)
}
\arguments{
\item{coords}{dataframe contains the spatial position of each cell. Column
names c("x", "y'). Must include row
names as integer from 1 to nrow(coords). This is automatically included in
the wSIR function prior to this function.}

\item{slices}{integer the number of slices along each of the two spatial
axes. The number of tiles will be
slices^2 since there are two spatial axes. E.g set slices = 4 to use
4^2 = 16 tiles. Default value is 3.}
}
\value{
output by itself is a matrix containing slice belonging for each
axis in long format. When used in
lapply(coords_split, spatialAllocator, slices) as in wSIR the output is
a dataframe with each cell's tile
allocation in the "coordinate" column.
}
\description{
This function allocates each cell to a tile based on a specified number of
tiles.
}
\keyword{internal}
