% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_lower_tri.R
\name{subset_lower_tri}
\alias{subset_lower_tri}
\title{subset_lower_tri}
\usage{
subset_lower_tri(m)
}
\arguments{
\item{m}{to fill}
}
\value{
to fill
}
\description{
This function is within the code for finding correlation of distances
}
\keyword{internal}
