% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_genes.R
\name{top_genes}
\alias{top_genes}
\title{top_genes}
\usage{
top_genes(WSIR, highest = 10)
}
\arguments{
\item{WSIR}{wsir object as output of wSIR function. To analyse a different DR method, ensure the 
slot named 'directions' contains the loadings as a matrix with the gene names as the rownames.}

\item{highest}{integer for how many of the top genes you would like to see. Recommend no more 
than 20 for ease of visualisation. Default is 10.}
}
\value{
List containing two slots. First is named "plot" and shows a barplot with the top genes 
and their corresponding loadings. Second is named "genes" and is a vector of the genes with highest
(in absolute value) loading in low-dimensional direction 1. Length is parameter highest.
}
\description{
A function to find and visualise the genes with the highest (in absolute value) loading in WSIR1. 
These genes contribute the most to the first low-dimensional direction.
}
\examples{
wsir_obj = wSIR(exprs = sample1_exprs, coords = sample1_coords) # create wsir object
top_genes_obj = top_genes(WSIR = wsir_obj, highest = 8) # create top genes object
top_genes_plot = top_genes_obj$plot # select plot
top_genes_plot # print plot

}
