% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_umap.R
\name{vis_umap}
\alias{vis_umap}
\title{vis_umap}
\usage{
vis_umap(
  exprs,
  WSIR,
  highest_genes,
  n_genes = length(names(highest_genes$genes))
)
}
\arguments{
\item{exprs}{matrix containing normalised gene expression data including n cells and p genes, dimension n * p.}

\item{WSIR}{wsir object that is output of wSIR function. If you wish to generate UMAP plots based on other DR methods, ensure
that the slot named "scores" in WSIR parameter contains the low-dimensional representation of exprs.}

\item{highest_genes}{output from top_genes function.}

\item{n_genes}{integer for the number of genes you would like to show. Default is the number of genes you selected in the
top_genes function.}
}
\value{
Grid of umap plots with n_genes number of plots. Each shows the cells in a UMAP generated on the low-dimensional gene
expression data, coloured by their value for each of the genes found by top_genes.
}
\description{
A function to generate UMAP plots on the low-dimensional embedding of the gene expression data. The points are coloured by
their value for the genes with highest (in absolute value) loading in WSIR1.
}
\examples{
wsir_obj = wSIR(exprs = sample1_exprs, coords = sample1_coords) # create wsir object
top_genes_obj = top_genes(WSIR = wsir_obj, highest = 4) # create top genes object
umap_plots = vis_umap(exprs = sample1_exprs, WSIR = wsir_obj, highest_genes = top_genes_obj, n_genes = 4)
umap_plots

}
