% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_wsir.R
\name{visualise_wsir}
\alias{visualise_wsir}
\title{visualise_wsir}
\usage{
visualise_wsir(coords, WSIR, dirs = 6, mincol = "blue", maxcol = "red")
}
\arguments{
\item{coords}{dataframe containing spatial positions of n cells in 2D space. Dimension n * 2. Column names must be c("x", "y").}

\item{WSIR}{wsir object as output of wSIR function. To analyse a different DR method, ensure the 
slot named 'directions' contains the loadings as a matrix with the gene names as the rownames. Must 
have used the same coords parameter as in coords parameter for this function.}

\item{dirs}{integer for how many of the low-dimensional directions you would like to visualise. Recommend no more 
than 10 for ease of visualisation. Default is 6.}

\item{mincol}{String for the colour of low values of low-dimensional directions. Personal choice for user, default is "blue".}

\item{maxcol}{String for the colour of high values of low-dimensional directions. Personal choice for user, default is "red".}
}
\value{
Grid of plots with dirs number of plots. Each shows the cells at their spatial positions 
coloured by their value for each of the first 'dirs' WSIR directions.
}
\description{
A function to easily visualise the low-dimensional gene expression data. This function plots
each cell at its true spatial coordinates, coloured by its value for WSIR1 / WSIR2 / ... . 
The plots give an intuition about what biological signals are contained in the WSIR directions.
}
\examples{
wsir_obj = wSIR(exprs = sample1_exprs, coords = sample1_coords) # create wsir object
vis_obj = visualise_wsir(coords = sample1_coords, WSIR = wsir_obj, dirs = 8) # create visualisations
vis_obj

}
