% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocator.R
\name{allocator}
\alias{allocator}
\title{Day hello}
\usage{
allocator(nrows, slices = 8)
}
\arguments{
\item{nrows}{no arguments}

\item{slices}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
# hello()

}
