% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_weight_matrix.R
\name{cells_weight_matrix}
\alias{cells_weight_matrix}
\title{Day hello}
\usage{
cells_weight_matrix(coords, labels, alpha = 1)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
