% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreWSIRParams.R
\name{exploreWSIRParams}
\alias{exploreWSIRParams}
\title{Day hello}
\usage{
exploreWSIRParams(
  X,
  coords,
  alphaVals = seq(from = -10, to = 10, by = 1),
  numSlices = 3:10,
  nRepeats = 1,
  ...
)
}
\arguments{
\item{X}{no arguments}

\item{coords}{to fill}

\item{alphaVals}{to fill}

\item{numSlices}{to fill}

\item{nRepeats}{to fill}

\item{...}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
#hello()

}
