% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreWSIRParams.R
\name{exploreWSIRParams}
\alias{exploreWSIRParams}
\title{exploreWSIRParams function}
\usage{
exploreWSIRParams(
  exprs,
  coords,
  samples = rep(1, nrow(coords)),
  alpha_vals = c(0, 1, 2, 4, 8, 12),
  slice_vals = c(3, 5, 7, 10, 15, 20),
  varThreshold = 0.95,
  maxDirections = 50,
  metric = "DC",
  nrep = 5,
  verbose = FALSE
)
}
\arguments{
\item{exprs}{matrix containing normalised gene expression data including n cells and p genes, dimension n * p.}

\item{coords}{dataframe containing spatial positions of n cells in 2D space. Dimension n * 2. Column names must be c("x", "y").}

\item{samples}{sample ID of each cell. In total, must have length equal to the number of cells. For example, if
your dataset has 10000 cells, the first 5000 from sample 1 and the remaining 5000 from sample 2, you would write
samples = c(rep(1, 5000), rep(2, 5000)) to specify that the first 5000 cells are sample 1 and the remaining are sample 2.
Default is that all cells are from sample 1. Sample IDs can be of any format: for the previous example, you could write
samples = c(rep("sample 1", 5000), rep("sample 2", 5000)), and the result would be the same.}

\item{alpha_vals}{vector of numbers as the values of parameter alpha to use in WSIR. 0 gives Sliced Inverse Regression
(SIR) implementation, and larger values represent stronger spatial correlation. Suggest to use integers for interpretability,
but can use non-integers. Values must be non-negative.}

\item{slice_vals}{vector of integers as the values of parameter slices to use in WSIR. Suggest maximum value in the vector to
be no more than around \eqn{\sqrt{n/20}}, as this upper bound ensures an average of at least 10 cells per tile in the training set.}

\item{varThreshold}{numeric proportion of variance in \code{t(X_H) \%*\% W \%*\% X_H} to retain. Must be between 0 and 1. Default is 0.95.
Select higher threshold to include more dimensions, lower threshold to include less dimensions.}

\item{maxDirections}{integer for the maximum number of directions to include in the low-dimenensional embedding. Default is 50.}

\item{metric}{evaluation metric to use for parameter tuning. String, either "DC" to use distance correlation or "CD" to use
correlation of distances. Default is "DC".}

\item{nrep}{integer for the number of train/test splits of the data to perform.}

\item{verbose}{logical If TRUE, prints the current values of slices and alpha as the tuning gets up to performing WSIR
with each value. If FALSE, then no progress updates. Default is FALSE.}
}
\value{
List with four slots, named "plot" and "message".
\enumerate{
\item "Plot" shows the average metric value across the nrep iterations for every combination of parameters slices and alpha.
Larger circles for a slices/alpha combination indicates better performance for that pair of values.
\item "message" tells you the parameter combination with highest metric value.
\item "best_alpha" returns the integer for the best alpha values among the values that were tested.
\item "best_slices" returns the integer for the best slices value among the values that were tested.
}
}
\description{
This function is used to select the optimal values for parameters slices and alpha in weighted sliced inverse regression
based on your provided gene expression data and corresponding spatial coordinates. For a given evaluation metric,
it will visualise the performance of WSIR with varying function parameters based on your data, and return the optimal pair.
This pair of slices and alpha can be used for your downstream tasks.
}
\examples{
data(MouseData)
explore_params = exploreWSIRParams(exprs = sample1_exprs,
  coords = sample1_coords,
  alpha_vals = c(0,2,4,8),
  slice_vals = c(3,6,10))
explore_params$plot
explore_params$message
best_alpha = explore_params$best_alpha
best_slices = explore_params$best_slices
wsir_obj = wSIR(X = sample1_exprs,
  coords = sample1_coords,
  optim_params = FALSE,
  alpha = best_alpha,
  slices = best_slices)

}
