% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_categorical.R
\name{sir_categorical}
\alias{sir_categorical}
\title{Day hello}
\usage{
sir_categorical(X, Y, directions = 50, W = NULL, ...)
}
\arguments{
\item{X}{no arguments}

\item{Y}{to fill}

\item{directions}{to fill}

\item{W}{to fill}

\item{...}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
#hello()

}
