% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_PCA.R
\name{sir_PCA}
\alias{sir_PCA}
\title{Day hello}
\usage{
sir_PCA(
  sliced_data,
  directions,
  W = diag(nrow(sliced_data)),
  varThreshold = 0.99
)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
