% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_univariate.R
\name{sir_univariate}
\alias{sir_univariate}
\title{Day hello}
\usage{
sir_univariate(
  X,
  Y,
  directions = 10,
  categorical = FALSE,
  slices = 10,
  alpha = 1,
  W
)
}
\arguments{
\item{X}{no arguments}

\item{Y}{to fill}

\item{directions}{to fill}

\item{categorical}{to fill}

\item{slices}{to fill}

\item{alpha}{to fill}

\item{W}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
# hello()

}
