% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicer.R
\name{slicer}
\alias{slicer}
\title{Day hello}
\usage{
slicer(X, Y, slices = 8, categorical = FALSE)
}
\arguments{
\item{X}{no arguments}

\item{Y}{to fill}

\item{slices}{to fill}

\item{categorical}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
#hello()

}
