% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_allocator2.R
\name{spatial_allocator2}
\alias{spatial_allocator2}
\title{Day hello}
\usage{
spatial_allocator2(coords, slices = 3)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
