% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sir_package.R
\name{weighted_sir_package}
\alias{weighted_sir_package}
\title{Day hello}
\usage{
weighted_sir_package(
  X,
  coords,
  slices = 8,
  directions = 10,
  W = diag(slices)/nrow(X),
  alpha = 1
)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
