% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wSIR-package.R
\docType{package}
\name{wSIR-package}
\alias{wSIR-package}
\title{wSIR: Weighted Sliced Inverse Regression (wSIR) for supervised dimension reduction of spatial transcriptomics and single cell gene expression data}
\description{
Weighted Sliced Inverse Regression (wSIR) is a supervised dimension reduction algorithm for spatial transcriptomics gene expression data. For a provided gene expression matrix and dataframe of each cell's spatial coordinates, wSIR creates a low-dimensional representation of the gene expression data that preserves the ability to predict spatial coordinates that was present in the gene expression data. Furthermore, wSIR provides interpretable loadings which allow for projection of new single-cell gene expression data into a low-dimensional space which preserves the spatial information present in the gene expression data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sydneybiox.github.io/wSIR}
  \item \url{https://sydneybiox.github.io/wSIR/}
  \item Report bugs at \url{https://github.com/sydneybiox/wSIR/issues}
}

}
\author{
\strong{Maintainer}: Max Woollard \email{mwoo5086@uni.sydney.edu.au} (\href{https://orcid.org/0009-0000-6319-6926}{ORCID}) [contributor]

Authors:
\itemize{
  \item Linh Nghiem (\href{https://orcid.org/0000-0003-2874-9067}{ORCID}) [contributor]
  \item Shila Ghazanfar \email{shazanfar@gmail.com} (\href{https://orcid.org/0000-0001-7861-6997}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Pratibha Panwar (\href{https://orcid.org/0000-0002-7437-7084}{ORCID}) [contributor]
}

}
\keyword{internal}
