% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wSIR.R
\name{wSIR}
\alias{wSIR}
\title{Day hello}
\usage{
wSIR(
  X,
  coords,
  groups = rep(1, ncol(nrow(coords))),
  slices = 8,
  weighted = TRUE,
  alpha = 1,
  maxDirections = 50,
  ...
)
}
\arguments{
\item{none}{no arguments}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
hello()

}
