% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wSIR.R
\name{wSIR}
\alias{wSIR}
\title{Day hello}
\usage{
wSIR(
  X,
  coords,
  groups = rep(1, nrow(coords)),
  slices = 8,
  weighted = TRUE,
  alpha = 1,
  maxDirections = 50,
  ...
)
}
\arguments{
\item{X}{no arguments}

\item{coords}{to fill}

\item{groups}{to fill}

\item{slices}{to fill}

\item{weighted}{to fill}

\item{alpha}{to fill}

\item{maxDirections}{to fill}

\item{...}{to fill}
}
\value{
prints hello world
}
\description{
This function says hello
}
\examples{
#hello()

}
