% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wSIROptimisation.R
\name{wSIROptimisation}
\alias{wSIROptimisation}
\title{wSIROptimisation}
\usage{
wSIROptimisation(
  exprs,
  coords,
  slices,
  alpha,
  maxDirections,
  varThreshold,
  metric,
  nrep = 3
)
}
\arguments{
\item{exprs}{matrix of normalised gene expression data for n genes across p cells.}

\item{coords}{dataframe containing spatial positions of n cells in 2D space. Dimension n * 2. Column names must be c("x", "y").}

\item{slices}{integer for number of slices on each axis of tissue. For example, slices = 4 creates 4 slices along each spatial axis, yielding 4^2 = 16 tiles.
Default is 8, suggested minimum of 3. Suggest to tune this parameter using exploreWSIRParams() function.}

\item{alpha}{integer to specify desired strength of spatial correlation. alpha = 0 gives SIR implementation. Larger values give stronger spatial correlations.
Default value is 4, at which weight matrix value for a pair of tiles is inversely proportional to the physical distance between them. Suggest to tune this
parameter using exploreWSIRParams() function.}

\item{maxDirections}{integer for the maximum number of directions to include in the low-dimenensional embedding. Default is 50.}

\item{varThreshold}{numeric proportion of variance in \code{t(X_H) \%*\% W \%*\% X_H} to retain. Must be between 0 and 1. Default is 0.95.
Select higher threshold to include more dimensions, lower threshold to include less dimensions.}

\item{metric}{evaluation metric to use for parameterr tuning. String, either "DC" to use distance correlation or "CD" to use
correlation of distances. Default is "DC".}

\item{nrep}{integer for the number of train/test splits of the data to perform.}
}
\value{
Average metric value for the given metric over each train/test split.
}
\description{
This function is used for parameter optimisation in WSIR. with provided parameter values, exprs and coords, this function splits the data into equal train/test
splits a given number of times (nrep times). It computes the selected evaluation metric (distance correlation or correlation of distances) on the current test
set and returns the average value for that metric over all train/test splits.
}
\keyword{internal}
